backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
openai
phonenumbers
python-dotenv

